refactor: remove whitespace changes and minor labels
