+++ api/mfa_test.go
@@ -81,27 +81,22 @@ func (ts *MFATestSuite) TestMFADisable() {
 func (ts *MFATestSuite) TestMFARecoveryCodeGeneration() {
 	const EXPECTED_NUM_OF_RECOVERY_CODES = 8
 
-	u, err := models.NewUser(ts.instanceID, "", "test1@example.com", "test", ts.Config.JWT.Aud, nil)
-	u.EnableMFA()
-
-	err = ts.API.db.Create(u)
-	require.NoError(ts.T(), err)
-	token, err := generateAccessToken(u, time.Second*time.Duration(ts.Config.JWT.Exp), ts.Config.JWT.Secret)
-	require.NoError(ts.T(), err)
-
-	user, err := models.FindUserByEmailAndAudience(ts.API.db, ts.instanceID, "test1@example.com", ts.Config.JWT.Aud)
+	user, err := models.FindUserByEmailAndAudience(ts.API.db, ts.instanceID, "test@example.com", ts.Config.JWT.Aud)
 	ts.Require().NoError(err)
+	require.NoError(ts.T(), user.EnableMFA(ts.API.db))
+
+	token, err := generateAccessToken(user, time.Second*time.Duration(ts.Config.JWT.Exp), ts.Config.JWT.Secret)
+	require.NoError(ts.T(), err)
 
 	w := httptest.NewRecorder()
 	req := httptest.NewRequest(http.MethodGet, fmt.Sprintf("/mfa/%s/generate_recovery_codes", user.ID), nil)
 	req.Header.Set("Authorization", fmt.Sprintf("Bearer %s", token))
 	ts.API.handler.ServeHTTP(w, req)
+	require.Equal(ts.T(), http.StatusOK, w.Code)
 
 	data := make(map[string]interface{})
-	require.Equal(ts.T(), http.StatusOK, w.Code)
 	require.NoError(ts.T(), json.NewDecoder(w.Body).Decode(&data))
 
-	recoveryCodes := data["RecoveryCodes"].([]interface{})
-	numCodes := len(recoveryCodes)
-	require.Equal(ts.T(), EXPECTED_NUM_OF_RECOVERY_CODES, numCodes)
+	recoveryCodes := data["recovery_codes"].([]interface{})
+	require.Equal(ts.T(), EXPECTED_NUM_OF_RECOVERY_CODES, len(recoveryCodes))
 }
