+++ models/recovery_code_test.go
@@ -1,27 +1,23 @@
-package api
+package models
 
 import (
 	"testing"
 
-	"github.com/gofrs/uuid"
 	"github.com/netlify/gotrue/conf"
 	"github.com/netlify/gotrue/storage"
 	"github.com/netlify/gotrue/storage/test"
-	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"github.com/stretchr/testify/suite"
 )
 
-
 type RecoveryCodeTestSuite struct {
-    suite.Suite
-    db *storage.Connection
+	suite.Suite
+	db *storage.Connection
 }
 
 func (ts *RecoveryCodeTestSuite) SetupTest() {
-    TruncateAll(ts.db)
+	TruncateAll(ts.db)
 }
-
 
 func TestRecoveryCode(t *testing.T) {
 	globalConfig, err := conf.LoadGlobal(modelsTestConfig)
